pyproject.toml
[project]
name = "pharmabolt-backend-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor-Ebuka",email = "v.ebukaagu2016@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (==0.115.8)",
    "uvicorn (==0.34.0)",
    "sqlalchemy (==2.0.38)",
    "pydantic (==2.10.6)",
    "psycopg2-binary (==2.9.10)",
    "python-dotenv (==1.0.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "click (==8.1.8)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "idna (==3.10)",
    "pydantic-core (==2.27.2)",
    "sniffio (==1.3.1)",
    "starlette (==0.45.3)",
    "typing-extensions (==4.12.2)",
    "alembic (>=1.14.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]